MyTunes: Initial Upload
File set for MyTunes class exercise. The version here is largely the
class version, with the following updates:

Added Track Number to track list. Rewrote HTML modal to add this input
field and changed code to be FORM tags.

Modal button event handler was updated to POST the track number to
new-track.php, via AJAX. the function, showTracks(), was updated to
include the track number in the list.

new-track.php was modified to upload the new track number as well.