MyTunes: Add New Album Feature
Added new features for the MyTunes app:

Add New Album: Each artist album box has a footer with a feature to add
an album. The "+ Add an Album" legend is dynamic, changing to an input
box set, allowing the user to input album name, release date and an
image file name. When submitted, the database is updated with this
information, the local "music" object in browser memory is updated, and
the HTML page is dynamically updated (DOM changed, rather than reloading
the page.)

The code also handles adding an album to an artist with no albums. In
this case, the "music" object has no "albums" nested object, so that has
to first be created before adding the new album data to the local
browser memory object.

Album images changed from BLOB's (images stored on DB) to file names.
This was done to work with this alternative image store method. Image
files still need to be loaded onto the server, into an "images" folder,
so a user would need access there with FTP.

Improved the album layouts (CSS changes).

Javascript now sorts the tracks by track number (rather than track_id)
before outputing the list to the page. This way, tracks can be entered
out of order, but will always display in track order.