Healdsburg Galleries (CS53.11 Project)
Initial commit for the Healdsburg Galleries Website. This dynamic site
is framed in the index.php file, and built using the healdsgal.js file.
get-galleries.php fetches the galleries MySQL database contents needed
for the site, and get-genres.php fetches a subset used to create a
select-option menu for a Gallery Genres filter. JSONphp is used as it
handles UTF-16 codes properly. Images used for the site are kept in an
images subfolder. Bootstrap files are used to add some of the responsive
design elements, as well as some features like the collapsible accordion
button-bars and some of the nav-tabs. The CSS file contains further
media-specific customizations, as well as styling and some bootstrap
overrides.

The MySQL database has tables for galleries' names, addresses, phone,
etc.; gallery descriptions; gallery logos (as directory links;) and
genres with a secondary table to allow many-to-many connections with the
galleries table. Also structured but not used are the contacts and
events tables.

The website can be found at www.healdsburggalleries.com (/test, until
promoted to the main directory.)