Healdsburg Galleries: Added Artwalk Brochure
Added printable Artwalk Brochure feature to the Art Walk section of the
site. Click the link (under the upcoming Art Walk date) and a new page
opens with a printable JPEG. This image has an overlay on the top with
title text and php-inserted date for the next Art Walk. The look mimics
a current printed brochure, but updates automatically each month.