Healdsburg Galleries: Add Print Artwalk
Added feature to Artwalk Brochure page: A "Print" button above the
brochure image that calls up the browser's printer prompt. A small
javascript bean loads as long as Javascript is turned on, that places
the "Print" feature on the page and hangs a "click " event handler on
it. The handler hides the Print button, performs a window.print(), and
then unhides the Print button.